+++ verigrad_rl/envs/__init__.py
@@ -2,7 +2,15 @@
 
 from verigrad_rl.envs.arithmetic import ArithmeticEnv
 from verigrad_rl.envs.base import Task, TextEnv
+from verigrad_rl.envs.biosafety import BioSafetyEnv
 from verigrad_rl.envs.safety_circuit import SafetyCircuitEnv
 from verigrad_rl.envs.string_transform import StringTransformEnv
 
-__all__ = ["ArithmeticEnv", "SafetyCircuitEnv", "StringTransformEnv", "Task", "TextEnv"]
+__all__ = [
+    "ArithmeticEnv",
+    "BioSafetyEnv",
+    "SafetyCircuitEnv",
+    "StringTransformEnv",
+    "Task",
+    "TextEnv",
+]